frontend/requirements.txt
gradio>=4.44
Pillow>=10.0
